Handle fork resolution in chain controller
This continues my efforts to draw a cleaner line between whether a block
is valid and whether it is committed. The block validator now only
validates a block and the chain controller decides whether to actually
commit it or not.
